Feat: Added auth.go main file. New constructors updated in main.go
